Moved parser to its own class and fix up key overloads
