+++ src/Tmds.Ssh/PrivateKeyCredential.cs
@@ -6,15 +6,12 @@ using System.Text;
 
 namespace Tmds.Ssh;
 
-public sealed partial class PrivateKeyCredential : Credential
+public sealed class PrivateKeyCredential : Credential
 {
     internal string FilePath { get; }
     internal byte[] Passphrase { get; }
 
-    public PrivateKeyCredential(string path) : this(path, null)
-    { }
-
-    public PrivateKeyCredential(string path, string? passphrase)
+    public PrivateKeyCredential(string path, string? passphrase = null)
     {
         FilePath = path ?? throw new ArgumentNullException(nameof(path));
         Passphrase = string.IsNullOrEmpty(passphrase) ? Array.Empty<byte>() : Encoding.UTF8.GetBytes(passphrase);
